Update the views to show the upload and network status whether or not there are more activities
